+++ lib/src/ui/pages/cupertino/slider.g.dart
@@ -1,0 +1,39 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'slider.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$configHash() => r'6ee9e28fd84c70c93f715d9560bafa180289b145';
+
+/// See also [Config].
+@ProviderFor(Config)
+final configProvider =
+    AutoDisposeNotifierProvider<Config, SliderConfig>.internal(
+  Config.new,
+  name: r'configProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$configHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$Config = AutoDisposeNotifier<SliderConfig>;
+String _$theValueHash() => r'76d994db7ff31813af6855f0337d045a38c346b2';
+
+/// See also [TheValue].
+@ProviderFor(TheValue)
+final theValueProvider = AutoDisposeNotifierProvider<TheValue, double>.internal(
+  TheValue.new,
+  name: r'theValueProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$theValueHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$TheValue = AutoDisposeNotifier<double>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member