+++ lib/src/ui/pages/cupertino/widgets.dart
@@ -1,0 +1,9 @@
+import 'package:dreambook/src/ui/pages/shared/shared_code_view.dart';
+
+import 'slider.dart';
+import 'switch.dart';
+
+final cupertinoItems = <CodeItem>[
+  switchItem,
+  sliderItem,
+];