+++ lib/src/ui/pages/shared/code_space/code_span.dart
@@ -1,4 +1,0 @@
-// class StaticCodes {
-//   static const material = "import 'package:flutter/material.dart';";
-//   static const widget = "import 'package:flutter/widget.dart';";
-// }